优化定时器算法，采用linux内核定时器一样的算法

+++ engine/src/libCoreCommon/core_connection_mgr.h
@@ -27,7 +27,7 @@ namespace core
 		bool				init(uint32_t nMaxConnectionCount);
 		bool				connect(const std::string& szHost, uint16_t nPort, uint32_t nType, const std::string& szContext, uint32_t nSendBufferSize, uint32_t nRecvBufferSize);
 		bool				listen(const std::string& szHost, uint16_t nPort, uint32_t nType, const std::string& szContext, uint32_t nSendBufferSize, uint32_t nRecvBufferSize);
-		void				update(int32_t nTime);
+		void				update(int64_t nTime);
 
 		void				broadcast(uint32_t nType, uint8_t nMessageType, const void* pData, uint16_t nSize, const std::vector<uint64_t>* vecExcludeID);
